[FE] feat #251: 유저 프로필이미지 업데이트 API 연결
