+++ frontend/src/services/my/api.ts
@@ -1,6 +1,18 @@
-import { IEmptyNotificationList } from '@/types/myFreeAlert';
-// import { APIResponse } from "@/types/model";
 import { axiosAuthInstance } from '@/apis/axiosInstance';
+import { APIResponse } from '@/types/model';
+import { IPageableReq, IMyFavoritCampListRes, IEmptyNotificationList } from '@/types/my';
+
+// 내가 찜한 캠핑장 조회
+export const fetchFavoriteCamps = async ({
+  pageable
+}: IPageableReq): Promise<IMyFavoritCampListRes> => {
+  const response = await axiosAuthInstance.get<APIResponse<IMyFavoritCampListRes>>(`/mypage/campsite`, {
+    params: {
+      ...pageable
+    },
+  });
+  return response.data.data;
+};
 
 // 빈자리 알림 조회 API 요청
 export const fetchMyAlerts = async (): Promise<IEmptyNotificationList> => {
