--- frontend/src/types/myPageTab.ts
@@ -1,2 +1,0 @@
-export type TabKey =
-'reservations' | 'reviews' | 'favorites' | 'alerts' | 'profile';
